removed warnings

+++ src/main/java/org/teagle/vcttool/control/RandomKey.java
@@ -4,6 +4,6 @@ public class RandomKey
 {
 	public static String randomKey()
 	{
-		return String.valueOf((int)(Math.random()*(double)(1<<30)));
+		return String.valueOf((int)(Math.random()*(1<<30)));
 	}
 }
